Constants.swift splitted in different files
