+++ xdrip/Calibration/Protocol/Calibrator.swift
@@ -495,10 +495,10 @@ extension Calibrator {
     ///     - bgReading : reading that needs to be updated - inout parameter to improve performance
     private func updateCalculatedValue(for bgReading:BgReading) {
         if (bgReading.calculatedValue < 10) {
-            bgReading.calculatedValue = Constants.CalibrationAlgorithms.bgReadingErrorValue
+            bgReading.calculatedValue = ConstantsCalibrationAlgorithms.bgReadingErrorValue
             bgReading.hideSlope = true
         } else {
-            bgReading.calculatedValue = min(Constants.CalibrationAlgorithms.maximumBgReadingCalculatedValue, max(Constants.CalibrationAlgorithms.minimumBgReadingCalculatedValue, bgReading.calculatedValue))
+            bgReading.calculatedValue = min(ConstantsCalibrationAlgorithms.maximumBgReadingCalculatedValue, max(ConstantsCalibrationAlgorithms.minimumBgReadingCalculatedValue, bgReading.calculatedValue))
         }
     }
     
@@ -517,11 +517,11 @@ extension Calibrator {
     ///     - withSensor : the currently active sensor, optional
     private func calculateAgeAdjustedRawValue(for bgReading:BgReading, withSensor sensor:Sensor?) {
         if let sensor = sensor {
-            let adjustfor:Double = Constants.CalibrationAlgorithms.ageAdjustmentTime - (bgReading.timeStamp.toMillisecondsAsDouble() - sensor.startDate.toMillisecondsAsDouble())
+            let adjustfor:Double = ConstantsCalibrationAlgorithms.ageAdjustmentTime - (bgReading.timeStamp.toMillisecondsAsDouble() - sensor.startDate.toMillisecondsAsDouble())
             if (adjustfor <= 0 || !ageAdjustMentNeeded) {
                 bgReading.ageAdjustedRawValue = bgReading.rawData
             } else {
-                bgReading.ageAdjustedRawValue = ((Constants.CalibrationAlgorithms.ageAdjustmentFactor * (adjustfor / Constants.CalibrationAlgorithms.ageAdjustmentTime)) * bgReading.rawData) + bgReading.rawData
+                bgReading.ageAdjustedRawValue = ((ConstantsCalibrationAlgorithms.ageAdjustmentFactor * (adjustfor / ConstantsCalibrationAlgorithms.ageAdjustmentTime)) * bgReading.rawData) + bgReading.rawData
             }
         } else {
             bgReading.ageAdjustedRawValue = bgReading.rawData
