+++ xdrip/Constants/ConstantsBGGraphBuilder.swift
@@ -1,0 +1,5 @@
+enum ConstantsBGGraphBuilder {
+    static let maxSlopeInMinutes = 21
+    static let defaultLowMarkInMgdl = 70.0
+    static let defaultHighMmarkInMgdl = 170.0
+}
