+++ xdrip/Constants/ConstantsBloodGlucose.swift
@@ -1,0 +1,7 @@
+enum ConstantsBloodGlucose {
+    static let mmollToMgdl = 18.01801801801802
+    static let mgDlToMmoll = 0.0555
+    static let libreMultiplier = 117.64705
+}
+
+
