+++ xdrip/Constants/ConstantsBluetoothPairing.swift
@@ -1,0 +1,7 @@
+/// constants related to Bluetooth pairing
+enum ConstantsBluetoothPairing {
+    
+    /// minimum time in seconds between two pairing notifications
+    static let minimumTimeBetweenTwoPairingNotificationsInSeconds = 30
+    
+}
