+++ xdrip/Constants/ConstantsCoreData.swift
@@ -1,0 +1,4 @@
+/// coredata specific constants
+enum ConstantsCoreData {
+    static let modelName = "xdrip"
+}
