+++ xdrip/Constants/ConstantsDefaultAlertTypeSettings.swift
@@ -1,0 +1,12 @@
+/// default values to be used when creating a new AlertType
+enum ConstantsDefaultAlertTypeSettings {
+    
+    static let enabled = true
+    static let name = Texts_Common.default0
+    static let overrideMute = false
+    static let snooze = true
+    static let snoozePeriod = Int16(60)
+    static let vibrate = true
+    static let soundName:String? = nil
+}
+
