+++ xdrip/Constants/ConstantsHomeView.swift
@@ -1,0 +1,12 @@
+/// constants for home view, ie first view
+enum ConstantsHomeView {
+    
+    /// how often to update the labels in the homeview (ie label with latest reading, minutes ago, etc..)
+    static let updateHomeViewIntervalInSeconds = 15.0
+    
+    /// info email adres, appears in licenseInfo
+    static let infoEmailAddress = "xdrip@proximus.be"
+    
+    /// application name, appears in licenseInfo as title
+    static let applicationName = "xDrip"
+}
