+++ xdrip/Constants/ConstantsLog.swift
@@ -1,0 +1,50 @@
+/// for use in OSLog
+enum ConstantsLog {
+    /// for use in OSLog
+    static let subSystem = "xDrip"
+    /// for use in OSLog
+    static let categoryBlueTooth = "bluetooth"
+    /// for use in cgm transmitter miaomiao
+    static let categoryCGMMiaoMiao = "cgmmiaomiao"
+    /// for use in cgm xdripg4
+    static let categoryCGMxDripG4 = "cgmxdripg4"
+    /// for use in firstview
+    static let categoryFirstView = "firstview"
+    /// calibration
+    static let calibration = "Calibration"
+    /// debuglogging
+    static let debuglogging = "xdripdebuglogging"
+    /// G5
+    static let categoryCGMG5 = "categoryCGMG5"
+    /// GNSEntry
+    static let categoryCGMGNSEntry = "categoryCGMGNSEntry"
+    /// Blucon
+    static let categoryBlucon = "categoryBlucon"
+    /// core data manager
+    static let categoryCoreDataManager = "categoryCoreDataManager"
+    /// application data bgreadings
+    static let categoryApplicationDataBgReadings = "categoryApplicationDataBgReadings"
+    /// application data calibrations
+    static let categoryApplicationDataCalibrations = "categoryApplicationDataCalibrations"
+    /// application data sensors
+    static let categoryApplicationDataSensors = "categoryApplicationDataSensors"
+    /// application data alerttypes
+    static let categoryApplicationDataAlertTypes = "categoryApplicationDataAlertTypes"
+    /// application data alertentries
+    static let categoryApplicationDataAlertEntries = "categoryApplicationDataAlertEntries"
+    /// nightscout uploader
+    static let categoryNightScoutUploadManager = "categoryNightScoutUploadManager"
+    /// nightscout follow
+    static let categoryNightScoutFollowManager = "categoryNightScoutFollowManager"
+    /// alertmanager
+    static let categoryAlertManager = "categoryAlertManager"
+    /// playsound
+    static let categoryPlaySound = "categoryPlaySound"
+    /// healthkit manager
+    static let categoryHealthKitManager = "categoryHealthKitManager"
+    /// SettingsViewHealthKitSettingsViewModel
+    static let categorySettingsViewHealthKitSettingsViewModel = "categorySettingsViewHealthKitSettingsViewModel"
+    /// dexcom share upload manager
+    static let categoryDexcomShareUploadManager = "categoryDexcomShareUploadManager"
+}
+
