+++ xdrip/Constants/ConstantsSuspensionPrevention.swift
@@ -1,0 +1,11 @@
+/// suspension prevention
+enum ConstantsSuspensionPrevention {
+    
+    /// name of the file that has the sound to play
+    static let soundFileName = "1-millisecond-of-silence.mp3"//20ms-of-silence.caf"
+    
+    /// how often to play the sound, in seconds
+    static let interval = 5
+}
+
+
