+++ xdrip/Core Data/classes/AlertType+CoreDataClass.swift
@@ -10,7 +10,7 @@ import CoreData
 /// - overrideMute : should alert make sound if the alert is muted
 /// - snooze : can the alert be snoozed from the home screen notification
 /// - vibrate : should the phone vibrate when the alert fires
-/// - soundname : the name of the sound as shown in the UI. Constants.Sounds defines available sounds. Per case there a string which is the soundname as shown in the UI and the filename of the sound in the Resources folder, seperated by backslash. Some special cases : an empty string means no sound needed. A nil value means default iOS sound. Any other value should be in the list defined in Constants.Sounds, otherwise the default xDrip sound will be used (see AlertManager.swift)
+/// - soundname : the name of the sound as shown in the UI. ConstantsSounds defines available sounds. Per case there a string which is the soundname as shown in the UI and the filename of the sound in the Resources folder, seperated by backslash. Some special cases : an empty string means no sound needed. A nil value means default iOS sound. Any other value should be in the list defined in ConstantsSounds, otherwise the default xDrip sound will be used (see AlertManager.swift)
 /// - alertEntries : the alertEntries in which this AlertType is used, optional
 public class AlertType: NSManagedObject {
     init(
